fixed return type of vm

+++ include/Arcvm.h
@@ -54,7 +54,7 @@ class Arcvm {
         Arcvm();
         ~Arcvm();
         void loadProgram(char*, size_t);
-        u64 run();
+        i32 run();
 
     private:
         u64 stack_pointer;
@@ -68,7 +68,7 @@ class Arcvm {
         u8 *program;
         size_t size;
 
-        u64 exit_code;
+        i32 exit_code;
 
         void execute();
